Renaming cuda.py to utils.py
