+++ python/triton/language/extra/cuda/__init__.py
@@ -1,5 +1,5 @@
 from . import libdevice
 
-from .cuda import (globaltimer, num_threads, num_warps, smid)
+from .utils import (globaltimer, num_threads, num_warps, smid)
 
 __all__ = ["libdevice", "globaltimer", "num_threads", "num_warps", "smid"]
